refactor: FileRepositoryUtil 에서 postIdFileUtil 분리
